+++ src/components/pages/components/RouteToLayout.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Redirect, Route } from 'react-router-dom';
+import { Extension } from '../../../AppCore';
 import { useIsAuthenticated } from '../../../hooks/useIsAuthenticated';
 import { PageConfig } from '../../../types/pagesConfig';
 import { DefaultRoutes } from '../../../types/routing';
@@ -10,6 +11,7 @@ interface RouteToLayoutProps {
   path: string;
   pageConfig: PageConfig;
   defaultRoutes: DefaultRoutes;
+  extensions?: Extension[];
 }
 
 const RouteToLayout: React.FC<RouteToLayoutProps> = (props) => {
@@ -39,6 +41,7 @@ const RouteToLayout: React.FC<RouteToLayoutProps> = (props) => {
           rows={props.pageConfig.rows}
           subPages={props.pageConfig.subPages}
           defaultRoutes={props.defaultRoutes}
+          extensions={props.extensions}
         />
       }
     />